requirements.txt
psycopg2-binary>=2.9
pytest>=7.0
pytest-django>=4.5
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
oauthlib==3.2.2
packaging==25.0
pip==21.2.4
pluggy==1.5.0
pycodestyle==2.13.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==58.1.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3===2.4.0
python-dotenv==1.1.0

